feat: Criando componente separado para a NavBar e importando no Header
